+++ TodoApi/Models/DayLog.cs
@@ -5,10 +5,10 @@ namespace TodoApi.Models
 {
     public class DayLog
     {
-        public long Id { get; set;}
+        public int Id { get; set;}
         public DateTime Date { get; set;}
         public bool isComplete { get; set;}
-        public long TodoLogId {get; set;}
+        public int TodoLogId {get; set;}
         public TodoLog TodoLog {get; set;}
     }
 }
