+++ TodoApi/Services/ITodoItemManager.cs
@@ -7,7 +7,7 @@ namespace TodoApi.Service
     public interface ITodoItemManager
     {
        void CreateTodoItem(TodoItem item, TargetItem targetItem);
-       TodoItem GetTodoItemById(long id);
+       TodoItem GetTodoItemById(int id);
        void DeleteTodoIterm(TodoItem item);
 
        List<TodoItem> GetAllDefaultTodoItems(TargetItem targetItem);
